+++ lib/entities/docs/commercio_doc.dart
@@ -1,3 +1,4 @@
+import 'package:commerciosdk/utils/utils.dart';
 import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
@@ -6,7 +7,7 @@ part 'commercio_doc.g.dart';
 
 /// Contains all the data related to a document that is sent to the chain when
 /// a user wants to share a document with another user.
-@JsonSerializable(explicitToJson: true)
+@JsonSerializable(explicitToJson: true, includeIfNull: false)
 class CommercioDoc extends Equatable {
   @JsonKey(name: 'sender')
   final String senderDid;
@@ -29,6 +30,7 @@ class CommercioDoc extends Equatable {
   @JsonKey(name: 'encryption_data')
   final CommercioDocEncryptionData encryptionData;
 
+  /// If [doSign] is specified then the field [checksum] must be also provided.
   @JsonKey(name: 'do_sign')
   final CommercioDoSign doSign;
 
@@ -41,10 +43,15 @@ class CommercioDoc extends Equatable {
     this.checksum,
     this.encryptionData,
     this.doSign,
-  })  : assert(senderDid != null),
-        assert(recipientDids != null && recipientDids.isNotEmpty),
-        assert(uuid != null),
-        assert(metadata != null);
+  })  : assert(senderDid != null && matchBech32Format(senderDid)),
+        assert(recipientDids != null &&
+            recipientDids.isNotEmpty &&
+            recipientDids.every((e) => matchBech32Format(e))),
+        assert(uuid != null && matchUuidv4(uuid)),
+        assert(metadata != null),
+        assert(contentUri == null ||
+            (contentUri != null && checkStringBytesLen(contentUri, 512))),
+        assert(_checksumMustBePresentIfDoSign(checksum, doSign));
 
   @override
   List<Object> get props {
@@ -66,7 +73,7 @@ class CommercioDoc extends Equatable {
   Map<String, dynamic> toJson() => _$CommercioDocToJson(this);
 }
 
-@JsonSerializable(explicitToJson: true)
+@JsonSerializable(explicitToJson: true, includeIfNull: false)
 class CommercioDocMetadata extends Equatable {
   @JsonKey(name: 'content_uri')
   final String contentUri;
@@ -79,10 +86,13 @@ class CommercioDocMetadata extends Equatable {
 
   CommercioDocMetadata({
     @required this.contentUri,
-    this.schemaType = '',
+    this.schemaType,
     this.schema,
-  })  : assert(contentUri != null),
-        assert(schemaType.isNotEmpty || schema != null);
+  })  : assert(contentUri != null && checkStringBytesLen(contentUri, 512)),
+        assert((schemaType != null &&
+                schemaType.isNotEmpty &&
+                checkStringBytesLen(schemaType, 512)) ||
+            schema != null);
 
   @override
   List<Object> get props {
@@ -95,7 +105,7 @@ class CommercioDocMetadata extends Equatable {
   Map<String, dynamic> toJson() => _$CommercioDocMetadataToJson(this);
 }
 
-@JsonSerializable(explicitToJson: true)
+@JsonSerializable(explicitToJson: true, includeIfNull: false)
 class CommercioDocMetadataSchema extends Equatable {
   @JsonKey(name: 'uri')
   final String uri;
@@ -106,8 +116,10 @@ class CommercioDocMetadataSchema extends Equatable {
   CommercioDocMetadataSchema({
     @required this.uri,
     @required this.version,
-  })  : assert(uri != null),
-        assert(version != null);
+  })  : assert(uri != null && uri.isNotEmpty && checkStringBytesLen(uri, 512)),
+        assert(version != null &&
+            uri.isNotEmpty &&
+            checkStringBytesLen(version, 32));
 
   @override
   List<Object> get props {
@@ -120,7 +132,7 @@ class CommercioDocMetadataSchema extends Equatable {
   Map<String, dynamic> toJson() => _$CommercioDocMetadataSchemaToJson(this);
 }
 
-@JsonSerializable(explicitToJson: true)
+@JsonSerializable(explicitToJson: true, includeIfNull: false)
 class CommercioDocChecksum extends Equatable {
   @JsonKey(name: 'value')
   final String value;
@@ -131,7 +143,7 @@ class CommercioDocChecksum extends Equatable {
   CommercioDocChecksum({
     @required this.value,
     @required this.algorithm,
-  })  : assert(value != null),
+  })  : assert(value != null && value.isNotEmpty),
         assert(algorithm != null);
 
   @override
@@ -165,15 +177,15 @@ enum CommercioDocChecksumAlgorithm {
   SHA512,
 }
 
-@JsonSerializable(explicitToJson: true)
+@JsonSerializable(explicitToJson: true, includeIfNull: false)
 class CommercioDocEncryptionData extends Equatable {
   @JsonKey(name: 'keys')
   final List<CommercioDocEncryptionDataKey> keys;
 
   @JsonKey(name: 'encrypted_data')
-  final List<String> encryptedData;
+  final Set<CommercioEncryptedData> encryptedData;
 
-  CommercioDocEncryptionData({
+  const CommercioDocEncryptionData({
     @required this.keys,
     @required this.encryptedData,
   })  : assert(keys != null),
@@ -190,7 +202,7 @@ class CommercioDocEncryptionData extends Equatable {
   Map<String, dynamic> toJson() => _$CommercioDocEncryptionDataToJson(this);
 }
 
-@JsonSerializable(explicitToJson: true)
+@JsonSerializable(explicitToJson: true, includeIfNull: false)
 class CommercioDocEncryptionDataKey extends Equatable {
   @JsonKey(name: 'recipient')
   final String recipientDid;
@@ -202,7 +214,7 @@ class CommercioDocEncryptionDataKey extends Equatable {
     @required this.recipientDid,
     @required this.value,
   })  : assert(recipientDid != null),
-        assert(value != null);
+        assert(value != null && checkStringBytesLen(value, 512));
 
   @override
   List<Object> get props {
@@ -215,7 +227,7 @@ class CommercioDocEncryptionDataKey extends Equatable {
   Map<String, dynamic> toJson() => _$CommercioDocEncryptionDataKeyToJson(this);
 }
 
-@JsonSerializable(explicitToJson: true)
+@JsonSerializable(explicitToJson: true, includeIfNull: false)
 class CommercioDoSign extends Equatable {
   @JsonKey(name: 'storage_uri')
   final String storageUri;
@@ -223,8 +235,8 @@ class CommercioDoSign extends Equatable {
   @JsonKey(name: 'signer_instance')
   final String signerIstance;
 
-  @JsonKey(name: 'sdn_data')
-  final List<CommercioSdnData> sdnData;
+  @JsonKey(name: 'sdn_data', includeIfNull: true)
+  final Set<CommercioSdnData> sdnData;
 
   @JsonKey(name: 'vcr_id')
   final String vcrId;
@@ -240,8 +252,10 @@ class CommercioDoSign extends Equatable {
     this.sdnData,
   })  : assert(storageUri != null),
         assert(signerIstance != null),
-        assert(vcrId != null),
-        assert(certificateProfile != null);
+        assert(vcrId != null && checkStringBytesLen(vcrId, 64)),
+        assert(certificateProfile != null &&
+            checkStringBytesLen(certificateProfile, 32)),
+        assert(_sdnDataNotEmpty(sdnData));
 
   @override
   List<Object> get props {
@@ -272,4 +286,77 @@ enum CommercioSdnData {
 
   @JsonValue('country')
   COUNTRY,
+}
+
+// For more information see:
+/// https://docs.commercio.network/x/docs/#supported-encrypted-data
+enum CommercioEncryptedData {
+  /// Special identifier, references the document's file contents. Means that
+  /// the `aes_key` has been used to encrypt a file exchanged by other means of
+  /// communication.
+  @JsonValue('content')
+  CONTENT,
+
+  /// The value of the field `content_uri` must be encrypted.
+  @JsonValue('content_uri')
+  CONTENT_URI,
+
+  /// The value of the field `content_uri` inside the `metadata` must be
+  /// encrypted.
+  @JsonValue('metadata.content_uri')
+  METADATA_CONTENT_URI,
+
+  /// The value of the field `uri` inside the `schema` inside `metadata` must be
+  /// encrypted.
+  @JsonValue('metadata.schema.uri')
+  METADATA_SCHEMA_URI,
+}
+
+extension CommercioEncryptedDataExt on CommercioEncryptedData {
+  /// Returns the string representation of the [EncryptedData] enum.
+  String get value {
+    switch (this) {
+      case CommercioEncryptedData.CONTENT:
+        return 'content';
+      case CommercioEncryptedData.CONTENT_URI:
+        return 'content_uri';
+      case CommercioEncryptedData.METADATA_CONTENT_URI:
+        return 'metadata.content_uri';
+      case CommercioEncryptedData.METADATA_SCHEMA_URI:
+        return 'metadata.schema.uri';
+      default:
+        return null;
+    }
+  }
+
+  /// Returns the [EncryptedData] that corresponds to the [value] or [null] if
+  /// the [value] is not valid.
+  static CommercioEncryptedData fromValue(String value) {
+    switch (value) {
+      case 'content':
+        return CommercioEncryptedData.CONTENT;
+      case 'content_uri':
+        return CommercioEncryptedData.CONTENT_URI;
+      case 'metadata.content_uri':
+        return CommercioEncryptedData.METADATA_CONTENT_URI;
+      case 'metadata.schema.uri':
+        return CommercioEncryptedData.METADATA_SCHEMA_URI;
+      default:
+        return null;
+    }
+  }
+}
+
+/// Return [true] if the fields [doSign] and [checksum] are both not null if the
+/// first is specified.
+bool _checksumMustBePresentIfDoSign(
+  CommercioDocChecksum checksum,
+  CommercioDoSign doSign,
+) {
+  return doSign != null ? checksum != null : true;
+}
+
+/// Returns [true] if [sdnData] is [null] or, if specified, must be not empty.
+bool _sdnDataNotEmpty(Set<CommercioSdnData> sdnData) {
+  return sdnData != null ? sdnData.isNotEmpty : true;
 }