+++ lib/entities/keys/rsa_keys.dart
@@ -19,10 +19,10 @@ class RSAPublicKey implements PublicKey {
 
   @override
   String getEncoded() {
-    var pubKeySequence = ASN1Sequence();
+    final pubKeySequence = ASN1Sequence();
     pubKeySequence.add(ASN1Integer(pubKey.modulus));
     pubKeySequence.add(ASN1Integer(pubKey.exponent));
-    var dataBase64 = base64.encode(pubKeySequence.encodedBytes);
+    final dataBase64 = base64.encode(pubKeySequence.encodedBytes);
     return '-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8A$dataBase64\n-----END PUBLIC KEY-----';
   }
 }
@@ -34,20 +34,20 @@ class RSAPrivateKey implements PrivateKey {
   RSAPrivateKey(this.secretKey);
 
   String encodePrivateKeyToPemPKCS1() {
-    var topLevel = ASN1Sequence();
+    final topLevel = ASN1Sequence();
 
-    var version = ASN1Integer(BigInt.from(0));
-    var modulus = ASN1Integer(secretKey.n);
-    var publicExponent = ASN1Integer(secretKey.exponent);
-    var privateExponent = ASN1Integer(secretKey.d);
-    var p = ASN1Integer(secretKey.p);
-    var q = ASN1Integer(secretKey.q);
-    var dP = secretKey.d % (secretKey.p - BigInt.from(1));
-    var exp1 = ASN1Integer(dP);
-    var dQ = secretKey.d % (secretKey.q - BigInt.from(1));
-    var exp2 = ASN1Integer(dQ);
-    var iQ = secretKey.q.modInverse(secretKey.p);
-    var co = ASN1Integer(iQ);
+    final version = ASN1Integer(BigInt.from(0));
+    final modulus = ASN1Integer(secretKey.n);
+    final publicExponent = ASN1Integer(secretKey.exponent);
+    final privateExponent = ASN1Integer(secretKey.d);
+    final p = ASN1Integer(secretKey.p);
+    final q = ASN1Integer(secretKey.q);
+    final dP = secretKey.d % (secretKey.p - BigInt.from(1));
+    final exp1 = ASN1Integer(dP);
+    final dQ = secretKey.d % (secretKey.q - BigInt.from(1));
+    final exp2 = ASN1Integer(dQ);
+    final iQ = secretKey.q.modInverse(secretKey.p);
+    final co = ASN1Integer(iQ);
 
     topLevel.add(version);
     topLevel.add(modulus);
@@ -59,7 +59,7 @@ class RSAPrivateKey implements PrivateKey {
     topLevel.add(exp2);
     topLevel.add(co);
 
-    var dataBase64 = base64.encode(topLevel.encodedBytes);
+    final dataBase64 = base64.encode(topLevel.encodedBytes);
 
     return '-----BEGIN PRIVATE KEY-----\r\n$dataBase64\r\n-----END PRIVATE KEY-----';
   }