+++ lib/entities/kyc/buy_membership.g.dart
@@ -10,6 +10,7 @@ BuyMembership _$BuyMembershipFromJson(Map<String, dynamic> json) {
   return BuyMembership(
     membershipType: json['membership_type'] as String,
     buyerDid: json['buyer'] as String,
+    tsp: json['tsp'] as String,
   );
 }
 
@@ -17,4 +18,5 @@ Map<String, dynamic> _$BuyMembershipToJson(BuyMembership instance) =>
     <String, dynamic>{
       'membership_type': instance.membershipType,
       'buyer': instance.buyerDid,
+      'tsp': instance.tsp,
     };