+++ lib/kyc/export.dart
@@ -1,3 +1,4 @@
 export 'buy_membership_helper.dart';
 export 'invite_user_helper.dart';
-export 'membership_helper.dart';
+export 'kyc_helper.dart';
+export 'reward_pool_deposit_helper.dart';